ingest/requirements.txt
confluent_kafka==1.8.2
PyYAML==6.0
tweepy==4.9.0
